Refactor phone number formatting to handle QP phone # formatting
This change refactors the `FormatPhoneNumber` function to include
support for handling phone numbers generated by Quickpass. The new
functionality checks if a contact is made by Quickpass based on the API
vendor ID. If so, it ensures that the phone number is in the correct
format (beginning with "+1") and removes any leading area code or
country code if the length is less than 7. The refactor handles Quickpass's poor formatting of phone numbers that breaks our phone number formatting.

This is because Quickpass adds a + to the start of all phone numbers, even if the phone number does not start with the country code. This makes the phone number invalid and breaks phone number formatting. This fix works around that.